Tests/LibWeb/Text/expected/intrinsic-width-of-inline-block-with-border-box-min-width.txt

width under min-content: 70
width under max-content: 50
FAIL: min-content width differs from max-content width
